Refactors EditTransaction and AddTransaction to use the same widget
