+++ lib/views/edit_transaction_screen.dart
@@ -1,10 +1,9 @@
-import 'package:financeplanner/extensions/extensions.dart';
 import 'package:financeplanner/middleware/middleware.dart';
 import 'package:financeplanner/models/app_state.dart';
 import 'package:financeplanner/models/models.dart';
+import 'package:financeplanner/views/widgets/transaction_form.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:redux/redux.dart';
 
@@ -12,59 +11,15 @@ class EditTransactionScreen extends StatefulWidget {
   final Store<AppState> store;
   final Transaction transaction;
 
-  EditTransactionScreen({Key key, this.store, this.transaction})
-      : super(key: key);
+  EditTransactionScreen({Key key, this.store, this.transaction}) : super(key: key);
 
   @override
   State<StatefulWidget> createState() {
-    return new EditTransactionState(transaction);
+    return new EditTransactionState();
   }
 }
 
 class EditTransactionState extends State<EditTransactionScreen> {
-  DateTime selectedDate = DateTime.now();
-
-  TextEditingController _descriptionController = new TextEditingController();
-  TextEditingController _amountController = TextEditingController();
-  TextEditingController _dateController = new TextEditingController();
-  TextEditingController _categoryController = new TextEditingController();
-
-  final FocusNode _amountFocus = FocusNode();
-  final FocusNode _descriptionFocus = FocusNode();
-  final FocusNode _categoryFocus = FocusNode();
-  final FocusNode _dateFocus = FocusNode();
-  final Transaction transaction;
-
-  final _formKey = GlobalKey<FormState>();
-
-  final _prefixMoneyRegex =
-      new RegExp(r'^-?(([1-9][0-9]*|0)(\,|\.)?)?([0-9]{1,2})?$');
-  String previousAmountText;
-  TextSelection previousAmountSelection;
-
-  EditTransactionState(this.transaction) {
-    _amountController.addListener(() {
-      final String currentValue = _amountController.text;
-
-      if (currentValue.length > 0 &&
-          _prefixMoneyRegex.matchAsPrefix(currentValue) == null) {
-        _amountController.value = TextEditingValue(
-            text: previousAmountText, selection: previousAmountSelection);
-      } else {
-        previousAmountText = currentValue;
-        previousAmountSelection = _amountController.selection;
-      }
-    });
-
-    previousAmountText = _amountController.text;
-    previousAmountSelection = _amountController.selection;
-
-    _setDate(transaction.date);
-    _amountController.text = transaction.amount.toString();
-    _descriptionController.text = transaction.description;
-    _categoryController.text = transaction.category;
-  }
-
   @override
   Widget build(BuildContext context) {
     return StoreProvider<AppState>(
@@ -73,177 +28,18 @@ class EditTransactionState extends State<EditTransactionScreen> {
         appBar: AppBar(
           title: Text("Edit Transaction"),
         ),
-        body: Form(
-          key: _formKey,
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              Padding(
-                child: TextFormField(
-                  controller: _dateController,
-                  focusNode: _dateFocus,
-                  onTap: () => _selectDate(context),
-                  keyboardType: TextInputType.number,
-                  readOnly: true,
-                  decoration: InputDecoration(
-                    border: OutlineInputBorder(),
-                    labelText: 'Date',
-                  ),
-                  validator: (text) {
-                    if (text == null || text.isEmpty) {
-                      return 'Date is required';
-                    }
-                    return null;
-                  },
-                ),
-                padding: const EdgeInsets.all(8),
-              ),
-              Padding(
-                child: TextFormField(
-                  controller: _amountController,
-                  keyboardType: TextInputType.number,
-                  textInputAction: TextInputAction.next,
-                  focusNode: _amountFocus,
-                  onFieldSubmitted: (term) {
-                    _fieldFocusChange(context, _amountFocus, _descriptionFocus);
-                  },
-                  decoration: InputDecoration(
-                    border: OutlineInputBorder(),
-                    labelText: 'Amount',
-                  ),
-                  validator: (text) {
-                    if (text == null || text.isEmpty) {
-                      return 'Amount is required';
-                    }
-
-                    if (!text.isMoney()) {
-                      return "Invalid amount";
-                    }
-
-                    return null;
-                  },
-                ),
-                padding: const EdgeInsets.all(8),
-              ),
-              Padding(
-                child: TextFormField(
-                  controller: _descriptionController,
-                  textInputAction: TextInputAction.next,
-                  focusNode: _descriptionFocus,
-                  onFieldSubmitted: (term) {
-                    _fieldFocusChange(
-                        context, _descriptionFocus, _categoryFocus);
-                  },
-                  decoration: InputDecoration(
-                    border: OutlineInputBorder(),
-                    labelText: 'Description',
-                  ),
-                  validator: (text) {
-                    if (text == null || text.isEmpty) {
-                      return 'Description is required';
-                    }
-                    return null;
-                  },
-                ),
-                padding: const EdgeInsets.all(8),
-              ),
-              Padding(
-                child: TextFormField(
-                  controller: _categoryController,
-                  textInputAction: TextInputAction.done,
-                  focusNode: _categoryFocus,
-                  onFieldSubmitted: (term) {
-                    submitAction();
-                  },
-                  decoration: InputDecoration(
-                    border: OutlineInputBorder(),
-                    labelText: 'Category',
-                  ),
-                  validator: (text) {
-                    if (text == null || text.isEmpty) {
-                      return 'Category is required';
-                    }
-                    return null;
-                  },
-                ),
-                padding: const EdgeInsets.all(8),
-              ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  StoreConnector<AppState, VoidCallback>(
-                    converter: (store) {
-                      return () {
-                        submitAction();
-                      };
-                    },
-                    builder: (context, callback) {
-                      return Expanded(
-                          child: Padding(
-                              padding: const EdgeInsets.all(8),
-                              child: new RaisedButton(
-                                onPressed: callback,
-                                child: Text("Save"),
-                              )));
-                    },
-                  ),
-                ],
-              ),
-            ],
-          ),
+        body: TransactionForm.filled(
+          transaction: widget.transaction,
+          onSuccess: submitAction,
+          submitText: "Edit",
         ),
       ),
     );
   }
 
-  void submitAction() {
-    if (_formKey.currentState.validate()) {
-      final Transaction editedTransaction = Transaction(
-        id: widget.transaction.id,
-        category: _categoryController.text,
-        description: _descriptionController.text,
-        amount: _amountController.text.parseMoney(),
-        dateTime: selectedDate,
-      );
-      widget.store.dispatch(editTransaction(editedTransaction));
+  void submitAction(Transaction transaction) {
+    widget.store.dispatch(editTransaction(transaction));
 
-      Navigator.pop(context, editedTransaction);
-    }
-  }
-
-  _fieldFocusChange(
-      BuildContext context, FocusNode currentFocus, FocusNode nextFocus) {
-    currentFocus.unfocus();
-    FocusScope.of(context).requestFocus(nextFocus);
-  }
-
-  Future<Null> _selectDate(BuildContext context) async {
-    final DateTime picked = await showDatePicker(
-        context: context,
-        initialDate: selectedDate,
-        firstDate: DateTime(1900),
-        lastDate: DateTime(2200));
-
-    if (picked != null && picked != selectedDate) {
-      setState(() {
-        _setDate(picked);
-      });
-    }
-
-    _fieldFocusChange(context, _dateFocus, _amountFocus);
-  }
-
-  void _setDate(DateTime date) {
-    selectedDate = date;
-    _dateController.text = date.toDateFormat();
-  }
-
-  @override
-  void dispose() {
-    _amountController.dispose();
-    _dateController.dispose();
-    _descriptionController.dispose();
-    _categoryController.dispose();
-    super.dispose();
+    Navigator.pop(context, transaction);
   }
 }
